# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api Version 1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks/{task}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete a task",
                "operationId": "a7406ac56cf85985bce27d5a943df31d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/task_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task deletion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "Update a task",
                "operationId": "7ff9601268a99f6b4774474e326b898b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/task_id"
                    }
                ],
                "requestBody": {
                    "description": "Update a task",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful task update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/complete": {
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "Mark a task as complete",
                "operationId": "70fc014c1c8591b4338727788cc6f004",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/task_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task completion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "JsonResponse": {
                "title": "Successful task deletion",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Success."
                    }
                },
                "type": "object"
            },
            "401": {
                "title": "Unauthenticated access",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "403": {
                "title": "Unauthorized access",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized.."
                    }
                },
                "type": "object"
            },
            "404": {
                "title": "Resource Not Found",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model [Model] 9aaf739d-5aeb-4117-ab3a-829e635411172"
                    }
                },
                "type": "object"
            },
            "422": {
                "title": "Validation error response",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UpdateTaskRequest": {
                "title": "Update Task Request",
                "description": "Request body for updating an existing task",
                "properties": {
                    "parent_id": {
                        "description": "Optional parent task ID",
                        "type": "string",
                        "example": "9aa8c264-462d-4a68-8c6d-09629cdb1127",
                        "nullable": true
                    },
                    "priority": {
                        "description": "Task priority level (1-5)",
                        "type": "integer",
                        "example": 3
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string",
                        "example": "Task Title"
                    },
                    "description": {
                        "description": "Optional task description",
                        "type": "string",
                        "example": "Task Description",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TaskResource": {
                "title": "Task Success Response",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "example": "9aa8c705-b08c-4d9a-ba1b-a8510a80bdac"
                            },
                            "parent_id": {
                                "type": "string",
                                "example": "9aa8c705-b63c-40ec-b77b-0e5f91da7a74",
                                "nullable": true
                            },
                            "status": {
                                "type": "integer",
                                "example": 0
                            },
                            "status_name": {
                                "type": "string",
                                "example": "Todo"
                            },
                            "priority": {
                                "type": "integer",
                                "example": 3
                            },
                            "title": {
                                "type": "string",
                                "example": "Task Title"
                            },
                            "description": {
                                "type": "string",
                                "example": "Task Description",
                                "nullable": true
                            },
                            "created_at": {
                                "type": "string",
                                "example": "2000-01-01 23:59:59"
                            },
                            "completed_at": {
                                "type": "string",
                                "example": null,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "task_id": {
                "name": "task",
                "in": "path",
                "description": "ID of the task to be updated",
                "required": true,
                "schema": {
                    "type": "string"
                },
                "example": "9aaaf739d-5aeb-4117-ab3a-829e635411172"
            },
...
        },
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Enter Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}